WIP Issue #83 Move compute_and_log_metrics to scoring
